+++ app/src/test/java/com/farmgame/entity/PlantsTest.java
@@ -1,8 +1,6 @@
 package com.farmgame.entity;
 
-import com.farmgame.entity.Plants;
-
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
